move jsKZG to util
