+++ packages/tx/package.json
@@ -58,7 +58,6 @@
     "@ethereumjs/rlp": "^5.0.2",
     "@ethereumjs/util": "^9.1.0",
     "ethereum-cryptography": "^3.0.0",
-    "micro-eth-signer": "^0.11.0",
     "trusted-setups": "github:paulmillr/trusted-setups"
   },
   "devDependencies": {
