+++ packages/tx/src/index.ts
@@ -21,6 +21,3 @@ export * from './types.js'
 
 // Utils
 export * from './util.js'
-
-// KZG Wrapper
-export * from './jsKZGWrapper.js'
