+++ packages/util/package.json
@@ -93,7 +93,8 @@
   },
   "dependencies": {
     "@ethereumjs/rlp": "^5.0.2",
-    "ethereum-cryptography": "^3.0.0"
+    "ethereum-cryptography": "^3.0.0",
+    "micro-eth-signer": "^0.11.0"
   },
   "devDependencies": {
     "kzg-wasm": "^0.4.0"
