+++ packages/util/src/fast.ts
